@@ -3,26 +3,21 @@ import PropTypes from "prop-types"
 
 export default class AuthorizeBtn extends React.Component {
   static propTypes = {
-    className: PropTypes.string
-  }
-
-  onClick =() => {
-    let { authActions, authSelectors } = this.props
-    let definitions = authSelectors.definitionsToAuthorize()
-
-    authActions.showDefinitions(definitions)
+    onClick: PropTypes.func,
+    isAuthorized: PropTypes.bool,
+    showPopup: PropTypes.bool,
+    getComponent: PropTypes.func.isRequired
   }
 
   render() {
-    let { authSelectors, getComponent } = this.props
+    let { isAuthorized, showPopup, onClick, getComponent } = this.props
+
     //must be moved out of button component
     const AuthorizationPopup = getComponent("authorizationPopup", true)
-    let showPopup = !!authSelectors.shownDefinitions()
-    let isAuthorized = !!authSelectors.authorized().size
 
     return (
       <div className="auth-wrapper">
-        <button className={isAuthorized ? "btn authorize locked" : "btn authorize unlocked"} onClick={ this.onClick }>
+        <button className={isAuthorized ? "btn authorize locked" : "btn authorize unlocked"} onClick={onClick}>
           <span>Authorize</span>
           <svg width="20" height="20">
             <use href={ isAuthorized ? "#locked" : "#unlocked" } xlinkHref={ isAuthorized ? "#locked" : "#unlocked" } />
@@ -31,13 +26,5 @@ export default class AuthorizeBtn extends React.Component {
       { showPopup && <AuthorizationPopup /> }
       </div>
     )
-  }
-
-
-  static propTypes = {
-    getComponent: PropTypes.func.isRequired,
-    authSelectors: PropTypes.object.isRequired,
-    errActions: PropTypes.object.isRequired,
-    authActions: PropTypes.object.isRequired,
   }
 }
