@@ -3,17 +3,19 @@ import { Map } from "immutable"
 import PropTypes from "prop-types"
 import ImPropTypes from "react-immutable-proptypes"
 import win from "core/window"
-import { getExtensions } from "core/utils"
+import { getExtensions, getCommonExtensions, numberToString } from "core/utils"
 
 export default class ParameterRow extends Component {
   static propTypes = {
     onChange: PropTypes.func.isRequired,
     param: PropTypes.object.isRequired,
+    rawParam: PropTypes.object.isRequired,
     getComponent: PropTypes.func.isRequired,
     fn: PropTypes.object.isRequired,
     isExecute: PropTypes.bool,
     onChangeConsumes: PropTypes.func.isRequired,
     specSelectors: PropTypes.object.isRequired,
+    specActions: PropTypes.object.isRequired,
     pathMethod: PropTypes.array.isRequired,
     getConfigs: PropTypes.func.isRequired,
     specPath: ImPropTypes.list.isRequired
@@ -22,60 +24,102 @@ export default class ParameterRow extends Component {
   constructor(props, context) {
     super(props, context)
 
-    let { specSelectors, pathMethod, param } = props
-    let defaultValue = param.get("default")
-    let parameter = specSelectors.getParameter(pathMethod, param.get("name"), param.get("in"))
-    let value = parameter ? parameter.get("value") : ""
-    if ( defaultValue !== undefined && value === undefined ) {
-      this.onChangeWrapper(defaultValue)
-    }
+    this.setDefaultValue()
   }
 
   componentWillReceiveProps(props) {
-    let { specSelectors, pathMethod, param } = props
+    let { specSelectors, pathMethod, rawParam } = props
     let { isOAS3 } = specSelectors
 
-    let example = param.get("example")
-    let defaultValue = param.get("default")
-    let parameter = specSelectors.getParameter(pathMethod, param.get("name"), param.get("in"))
+    let parameterWithMeta = specSelectors.parameterWithMetaByIdentity(pathMethod, rawParam) || new Map()
+    // fallback, if the meta lookup fails
+    parameterWithMeta = parameterWithMeta.isEmpty() ? rawParam : parameterWithMeta
+
     let enumValue
 
     if(isOAS3()) {
-      let schema = param.get("schema") || Map()
+      let schema = parameterWithMeta.get("schema") || Map()
       enumValue = schema.get("enum")
     } else {
-      enumValue = parameter ? parameter.get("enum") : undefined
+      enumValue = parameterWithMeta ? parameterWithMeta.get("enum") : undefined
     }
-    let paramValue = parameter ? parameter.get("value") : undefined
+    let paramValue = parameterWithMeta ? parameterWithMeta.get("value") : undefined
 
     let value
 
     if ( paramValue !== undefined ) {
       value = paramValue
-    } else if ( example !== undefined ) {
-      value = example
-    } else if ( defaultValue !== undefined) {
-      value = defaultValue
-    } else if ( param.get("required") && enumValue && enumValue.size ) {
+    } else if ( rawParam.get("required") && enumValue && enumValue.size ) {
       value = enumValue.first()
     }
 
-    if ( value !== undefined ) {
-      this.onChangeWrapper(value)
+    if ( value !== undefined && value !== paramValue ) {
+      this.onChangeWrapper(numberToString(value))
+    }
+
+    this.setDefaultValue()
+  }
+
+  onChangeWrapper = (value, isXml = false) => {
+    let { onChange, rawParam } = this.props
+    let valueForUpstream
+    
+    // Coerce empty strings and empty Immutable objects to null
+    if(value === "" || (value && value.size === 0)) {
+      valueForUpstream = null
+    } else {
+      valueForUpstream = value
+    }
+
+    return onChange(rawParam, valueForUpstream, isXml)
+  }
+
+  onChangeIncludeEmpty = (newValue) => {
+    let { specActions, param, pathMethod } = this.props
+    const paramName = param.get("name")
+    const paramIn = param.get("in")
+    return specActions.updateEmptyParamInclusion(pathMethod, paramName, paramIn, newValue)
+  }
+
+  setDefaultValue = () => {
+    let { specSelectors, pathMethod, rawParam } = this.props
+
+    let paramWithMeta = specSelectors.parameterWithMetaByIdentity(pathMethod, rawParam)
+
+
+    if (!paramWithMeta || paramWithMeta.get("value") !== undefined) {
+      return
+    }
+
+    if( paramWithMeta.get("in") !== "body" ) {
+      let newValue
+
+      if (specSelectors.isSwagger2()) {
+        newValue = paramWithMeta.get("x-example")
+          || paramWithMeta.getIn(["default"])
+          || paramWithMeta.getIn(["schema", "example"])
+          || paramWithMeta.getIn(["schema", "default"])
+      } else if (specSelectors.isOAS3()) {
+        newValue = paramWithMeta.get("example")
+          || paramWithMeta.getIn(["schema", "example"])
+          || paramWithMeta.getIn(["schema", "default"])
+      }
+      if(newValue !== undefined) {
+        this.onChangeWrapper(numberToString(newValue))
+      }
     }
   }
 
-  onChangeWrapper = (value) => {
-    let { onChange, param } = this.props
-    return onChange(param, value)
-  }
-
   render() {
-    let {param, onChange, getComponent, getConfigs, isExecute, fn, onChangeConsumes, specSelectors, pathMethod, specPath} = this.props
+    let {param, rawParam, getComponent, getConfigs, isExecute, fn, onChangeConsumes, specSelectors, pathMethod, specPath} = this.props
 
     let { isOAS3 } = specSelectors
 
-    const { showExtensions } = getConfigs()
+    const { showExtensions, showCommonExtensions } = getConfigs()
+
+    if(!param) {
+      param = rawParam
+    }
 
     // const onChangeWrapper = (value) => onChange(param, value)
     const JsonSchemaForm = getComponent("JsonSchemaForm")
@@ -87,7 +131,7 @@ export default class ParameterRow extends Component {
                    param={param}
                    consumes={ specSelectors.operationConsumes(pathMethod) }
                    consumesValue={ specSelectors.contentTypeValues(pathMethod).get("requestContentType") }
-                   onChange={onChange}
+                   onChange={this.onChangeWrapper}
                    onChangeConsumes={onChangeConsumes}
                    isExecute={ isExecute }
                    specSelectors={ specSelectors }
@@ -97,78 +141,84 @@ export default class ParameterRow extends Component {
     const ModelExample = getComponent("modelExample")
     const Markdown = getComponent("Markdown")
     const ParameterExt = getComponent("ParameterExt")
+    const ParameterIncludeEmpty = getComponent("ParameterIncludeEmpty")
 
-    let schema = param.get("schema")
-    let type = isOAS3 && isOAS3() ? param.getIn(["schema", "type"]) : param.get("type")
+    let paramWithMeta = specSelectors.parameterWithMetaByIdentity(pathMethod, rawParam)
+    let format = param.get("format")
+    let schema = isOAS3 && isOAS3() ? param.get("schema") : param
+    let type = schema.get("type")
     let isFormData = inType === "formData"
     let isFormDataSupported = "FormData" in win
     let required = param.get("required")
-    let itemType = param.getIn(isOAS3 && isOAS3() ? ["schema", "items", "type"] : ["items", "type"])
-    let parameter = specSelectors.getParameter(pathMethod, param.get("name"), param.get("in"))
-    let value = parameter ? parameter.get("value") : ""
-    let extensions = getExtensions(param)
+    let itemType = schema.getIn(["items", "type"])
 
+    let value = paramWithMeta ? paramWithMeta.get("value") : ""
+    let commonExt = showCommonExtensions ? getCommonExtensions(param) : null
+    let extensions = showExtensions ? getExtensions(param) : null
 
     let paramItems // undefined
-    let paramItemsEnum // undefined
-    let isDisplayParamItemsEnum = false
+    let paramEnum // undefined
+    let paramDefaultValue // undefined
+    let paramExample // undefined
+    let isDisplayParamEnum = false
+
     if ( param !== undefined ) {
-      paramItems = param.get("items")
+      paramItems = schema.get("items")
     }
-    if ( paramItems !== undefined ) {
-      paramItemsEnum = param.get("items").get("enum")
+
+    if (paramItems !== undefined) {
+      paramEnum = paramItems.get("enum")
+      paramDefaultValue = paramItems.get("default")
+    } else {
+      paramEnum = schema.get("enum")
     }
-    if ( paramItemsEnum !== undefined ) {
-      if (paramItemsEnum.size > 0) {
-        isDisplayParamItemsEnum = true
-      }
+
+    if ( paramEnum !== undefined && paramEnum.size > 0) {
+      isDisplayParamEnum = true
     }
 
     // Default and Example Value for readonly doc
-    let paramDefaultValue // undefined
-    let paramExample // undefined
     if ( param !== undefined ) {
-      paramDefaultValue = param.get("default")
+      paramDefaultValue = schema.get("default")
       paramExample = param.get("example")
-    }
-
-    if (isDisplayParamItemsEnum) { // if we have an array, default value is in "items"
-      paramDefaultValue = paramItems.get("default")
+      if (paramExample === undefined) {
+        paramExample = param.get("x-example")
+      }
     }
 
     return (
-      <tr>
+      <tr data-param-name={param.get("name")} data-param-in={param.get("in")}>
         <td className="col parameters-col_name">
           <div className={required ? "parameter__name required" : "parameter__name"}>
             { param.get("name") }
             { !required ? null : <span style={{color: "red"}}>&nbsp;*</span> }
           </div>
-          <div className="parameter__type">{ type } { itemType && `[${itemType}]` }</div>
+          <div className="parameter__type">
+            { type }
+            { itemType && `[${itemType}]` }
+            { format && <span className="prop-format">(${format})</span>}
+          </div>
           <div className="parameter__deprecated">
             { isOAS3 && isOAS3() && param.get("deprecated") ? "deprecated": null }
           </div>
           <div className="parameter__in">({ param.get("in") })</div>
+          { !showCommonExtensions || !commonExt.size ? null : commonExt.map((v, key) => <ParameterExt key={`${key}-${v}`} xKey={key} xVal={v} /> )}
           { !showExtensions || !extensions.size ? null : extensions.map((v, key) => <ParameterExt key={`${key}-${v}`} xKey={key} xVal={v} /> )}
         </td>
 
         <td className="col parameters-col_description">
-          <Markdown source={ param.get("description") }/>
+          { param.get("description") ? <Markdown source={ param.get("description") }/> : null }
 
-          { (bodyParam || !isExecute) && isDisplayParamItemsEnum ?
-            <Markdown source={
-                "<i>Available values</i>: " + paramItemsEnum.map(function(item) {
+          { (bodyParam || !isExecute) && isDisplayParamEnum ?
+            <Markdown className="parameter__enum" source={
+                "<i>Available values</i> : " + paramEnum.map(function(item) {
                     return item
                   }).toArray().join(", ")}/>
             : null
           }
 
           { (bodyParam || !isExecute) && paramDefaultValue !== undefined ?
-            <Markdown source={"<i>Default value</i>: " + paramDefaultValue}/>
-            : null
-          }
-
-          { (bodyParam || !isExecute) && paramExample !== undefined ?
-            <Markdown source={"<i>Example</i>: " + paramExample}/>
+            <Markdown className="parameter__default" source={"<i>Default value</i> : " + paramDefaultValue}/>
             : null
           }
 
@@ -181,8 +231,8 @@ export default class ParameterRow extends Component {
                               required={ required }
                               description={param.get("description") ? `${param.get("name")} - ${param.get("description")}` : `${param.get("name")}`}
                               onChange={ this.onChangeWrapper }
-                              errors={ param.get("errors") }
-                              schema={ isOAS3 && isOAS3() ? param.get("schema") : param }/>
+                              errors={ paramWithMeta.get("errors") }
+                              schema={ schema }/>
           }
 
 
@@ -192,9 +242,19 @@ export default class ParameterRow extends Component {
                                                 getConfigs={ getConfigs }
                                                 isExecute={ isExecute }
                                                 specSelectors={ specSelectors }
-                                                schema={ schema }
+                                                schema={ param.get("schema") }
                                                 example={ bodyParam }/>
               : null
+          }
+
+          {
+            !bodyParam && isExecute ? 
+            <ParameterIncludeEmpty
+              onChange={this.onChangeIncludeEmpty}
+              isIncluded={specSelectors.parameterInclusionSettingFor(pathMethod, param.get("name"), param.get("in"))}
+              isDisabled={value && value.size !== 0}
+              param={param} /> 
+            : null
           }
 
         </td>
