@@ -1,3 +1,5 @@
+import get from "lodash/get"
+
 export const updateSpec = (ori, {specActions}) => (...args) => {
   ori(...args)
   specActions.parseToJson(...args)
@@ -5,7 +7,24 @@ export const updateSpec = (ori, {specActions}) => (...args) => {
 
 export const updateJsonSpec = (ori, {specActions}) => (...args) => {
   ori(...args)
-  specActions.resolveSpec(...args)
+
+  specActions.invalidateResolvedSubtreeCache()
+
+  // Trigger resolution of any path-level $refs.
+  const [json] = args
+  const pathItems = get(json, ["paths"]) || {}
+  const pathItemKeys = Object.keys(pathItems)
+
+  pathItemKeys.forEach(k => {
+    const val = get(pathItems, [k])
+
+    if(val.$ref) {
+      specActions.requestResolvedSubtree(["paths", k])
+    }
+  })
+
+  // Trigger resolution of any securitySchemes-level $refs.
+  specActions.requestResolvedSubtree(["components", "securitySchemes"])
 }
 
 // Log the request ( just for debugging, shouldn't affect prod )
